# src/DbBackupManager.Web/OpenApi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "DbBackupManager.Web | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/storage-targets": {
      "get": {
        "tags": [
          "StorageTargets"
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StorageTargetResponse"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminCookie": [ ]
          }
        ]
      },
      "post": {
        "tags": [
          "StorageTargets"
        ],
        "parameters": [
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "从 GET /api/v1/auth/csrf 获取，并与当前身份和配套 Antiforgery Cookie 一起使用。身份变化后必须重新获取。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStorageTargetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStorageTargetRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStorageTargetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageTargetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminCookie": [ ]
          }
        ]
      }
    },
    "/api/v1/storage-targets/{id}": {
      "put": {
        "tags": [
          "StorageTargets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "从 GET /api/v1/auth/csrf 获取，并与当前身份和配套 Antiforgery Cookie 一起使用。身份变化后必须重新获取。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStorageTargetRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStorageTargetRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SaveStorageTargetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageTargetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminCookie": [ ]
          }
        ]
      }
    },
    "/api/v1/storage-targets/{id}/enabled": {
      "post": {
        "tags": [
          "StorageTargets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "从 GET /api/v1/auth/csrf 获取，并与当前身份和配套 Antiforgery Cookie 一起使用。身份变化后必须重新获取。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetStorageTargetEnabledRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetStorageTargetEnabledRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetStorageTargetEnabledRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageTargetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminCookie": [ ]
          }
        ]
      }
    },
    "/api/v1/storage-targets/{id}/probe": {
      "post": {
        "tags": [
          "StorageTargets"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-CSRF-TOKEN",
            "in": "header",
            "description": "从 GET /api/v1/auth/csrf 获取，并与当前身份和配套 Antiforgery Cookie 一起使用。身份变化后必须重新获取。",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageTargetVersionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StorageTargetVersionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StorageTargetVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StorageTargetResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminCookie": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveStorageTargetRequest": {
        "required": [
          "name",
          "protocol",
          "host",
          "basePath",
          "credentialId",
          "isEnabled"
        ],
        "type": "object",
        "properties": {
          "version": {
            "maxLength": 16,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "name": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/StorageProtocolValue"
          },
          "host": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "port": {
            "maximum": 65535,
            "minimum": 1,
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "basePath": {
            "maxLength": 2048,
            "minLength": 0,
            "type": "string"
          },
          "credentialId": {
            "type": "string",
            "format": "uuid"
          },
          "fingerprint": {
            "maxLength": 500,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          },
          "isEnabled": {
            "type": "boolean"
          }
        }
      },
...
      "SetStorageTargetEnabledRequest": {
        "required": [
          "version",
          "isEnabled"
        ],
        "type": "object",
        "properties": {
          "version": {
            "maxLength": 16,
            "minLength": 0,
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          }
        }
      },
...
      "StorageProtocolValue": {
        "type": "integer"
      },
      "StorageTargetResponse": {
        "required": [
          "id",
          "settings",
          "version"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "settings": {
            "$ref": "#/components/schemas/StorageTargetSettingsResponse"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "StorageTargetSettingsResponse": {
        "required": [
          "name",
          "protocol",
          "host",
          "port",
          "basePath",
          "credentialId",
          "fingerprint",
          "isEnabled"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "protocol": {
            "$ref": "#/components/schemas/StorageProtocolValue"
          },
          "host": {
            "type": "string"
          },
          "port": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "basePath": {
            "type": "string"
          },
          "credentialId": {
            "type": "string",
            "format": "uuid"
          },
          "fingerprint": {
            "type": [
              "null",
              "string"
            ]
          },
          "isEnabled": {
            "type": "boolean"
          }
        }
      },
      "StorageTargetVersionRequest": {
        "required": [
          "version"
        ],
        "type": "object",
        "properties": {
          "version": {
            "maxLength": 16,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "AdminCookie": {
        "type": "apiKey",
        "description": "管理员登录后由同源 Web 主机签发的安全 Cookie。",
        "name": "DbBackupManager.Admin",
        "in": "cookie"
      }
    }
  },
...
}